Change default value of senttxstore to false
Currently this data is stored in plaintext on disk, which is less than ideal.
Users should explicitly opt-in to this feature. In the future, we plan to encrypt
the senttxstore. When that happens, we could default this feature to on again,
as long as the user has specified a PIN/password for the SilentDragon installation.