Fix autostart of hushd
Since hushd is no longer a script but a real binary, our code to autostart hushd no longer
has any arguments. But QT gets angry with an empty list of arguments, because it
splits arguments into an array on space characters. So this bug ended up tring to start hush
via ./hushd "" instead of ./hushd, which creates an "Ooops" error from hushd, which thinks you
are trying to use hush-cli.

Now we provide two arguments, with default values, that allows the rest of the code to work correctly.